error handling iniciado

+++ src/app/use-cases/CustomerDemand/delete-customer-demand.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { CustomerDemandsRepository } from '@app/repositories/CustomerDemands/customer-demands-repository'
+import { EntityIdNotProvidedError } from '../models/errors/entityIdNotProvided'
 
 interface DeleteCustomerDemandRequest {
   customerDemandId: string
@@ -11,6 +12,10 @@ export class DeleteCustomerDemand {
   constructor(private customerDemandRepository: CustomerDemandsRepository) {}
 
   async execute(request: DeleteCustomerDemandRequest): Promise<void> {
+    if (!request.customerId) {
+      throw new EntityIdNotProvidedError('Customer')
+    }
+
     await this.customerDemandRepository.delete(
       request.customerDemandId,
       request.customerId,
