+++ src/app/use-cases/Organization/delete-organization.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { OrganizationRepository } from '@app/repositories/Organization/organization-repository'
+import { EntityIdNotProvidedError } from '../models/errors/entityIdNotProvided'
 
 @Injectable()
 export class DeleteOrganization {
@@ -7,6 +8,10 @@ export class DeleteOrganization {
 
   async execute(request: string): Promise<void> {
     const organizationId = request
+
+    if (!organizationId) {
+      throw new EntityIdNotProvidedError('Organization')
+    }
 
     await this.organizationsRepository.delete(organizationId)
   }
