+++ src/app/use-cases/User/delete-user.ts
@@ -1,5 +1,6 @@
 import { UsersRepository } from '@app/repositories/User/users-repository'
 import { Injectable } from '@nestjs/common'
+import { EntityIdNotProvidedError } from '../models/errors/entityIdNotProvided'
 
 @Injectable()
 export class DeleteUser {
@@ -7,6 +8,10 @@ export class DeleteUser {
 
   async execute(request: string): Promise<void> {
     const userId = request
+
+    if (!userId) {
+      throw new EntityIdNotProvidedError('User')
+    }
 
     await this.usersRepository.deleteUser(userId)
   }
